main.py, html.py 파일 합침
